[Lagrangian.Solver] Move GenericConstraintProblem in its own files
