+++ Sofa/Component/Constraint/Lagrangian/Solver/src/sofa/component/constraint/lagrangian/solver/GenericConstraintSolver.h
@@ -21,11 +21,11 @@
 ******************************************************************************/
 #pragma once
 #include <sofa/component/constraint/lagrangian/solver/config.h>
+#include <sofa/component/constraint/lagrangian/solver/GenericConstraintProblem.h>
 
 #include <sofa/component/constraint/lagrangian/solver/ConstraintSolverImpl.h>
 #include <sofa/core/behavior/BaseConstraintCorrection.h>
 #include <sofa/core/behavior/BaseConstraint.h>
-#include <sofa/linearalgebra/SparseMatrix.h>
 #include <sofa/helper/map.h>
 
 #include <sofa/simulation/CpuTask.h>
@@ -33,62 +33,6 @@
 
 namespace sofa::component::constraint::lagrangian::solver
 {
-
-class GenericConstraintSolver;
-
-class SOFA_COMPONENT_CONSTRAINT_LAGRANGIAN_SOLVER_API GenericConstraintProblem : public ConstraintProblem
-{
-public:
-    sofa::linearalgebra::FullVector<SReal> _d;
-    std::vector<core::behavior::ConstraintResolution*> constraintsResolutions;
-    bool scaleTolerance, allVerified;
-    SReal sor;
-    SReal sceneTime;
-    SReal currentError;
-    int currentIterations;
-
-    // For unbuilt version :
-    sofa::linearalgebra::SparseMatrix<SReal> Wdiag;
-    std::list<unsigned int> constraints_sequence;
-    bool change_sequence;
-
-    typedef std::vector< core::behavior::BaseConstraintCorrection* > ConstraintCorrections;
-    typedef std::vector< core::behavior::BaseConstraintCorrection* >::iterator ConstraintCorrectionIterator;
-
-    std::vector< ConstraintCorrections > cclist_elems;
-
-
-    GenericConstraintProblem() : scaleTolerance(true), allVerified(false), sor(1.0)
-      , sceneTime(0.0), currentError(0.0), currentIterations(0)
-      , change_sequence(false) {}
-    ~GenericConstraintProblem() override { freeConstraintResolutions(); }
-
-    void clear(int nbConstraints) override;
-    void freeConstraintResolutions();
-    void solveTimed(SReal tol, int maxIt, SReal timeout) override;
-
-    /// Projective Gauss Seidel method building the compliance matrix
-    void gaussSeidel(SReal timeout=0, GenericConstraintSolver* solver = nullptr);
-    /// Projective Gauss Seidel unbuilt method
-    void unbuiltGaussSeidel(SReal timeout=0, GenericConstraintSolver* solver = nullptr);
-    /// Method from:
-    /// A nonsmooth nonlinear conjugate gradient method for interactive contact force problems
-    /// - 2010, Silcowitz, Morten and Niebe, Sarah and Erleben, Kenny
-    void NNCG(GenericConstraintSolver* solver = nullptr, int iterationNewton = 1);
-
-    void gaussSeidel_increment(bool measureError, SReal *dfree, SReal *force, SReal **w, SReal tol, SReal *d, int dim, bool& constraintsAreVerified, SReal& error, sofa::type::vector<SReal>& tabErrors) const;
-    void result_output(GenericConstraintSolver* solver, SReal *force, SReal error, int iterCount, bool convergence);
-
-    int getNumConstraints();
-    int getNumConstraintGroups();
-
-protected:
-    sofa::linearalgebra::FullVector<SReal> m_lam;
-    sofa::linearalgebra::FullVector<SReal> m_deltaF;
-    sofa::linearalgebra::FullVector<SReal> m_deltaF_new;
-    sofa::linearalgebra::FullVector<SReal> m_p;
-
-};
 
 class SOFA_COMPONENT_CONSTRAINT_LAGRANGIAN_SOLVER_API GenericConstraintSolver : public ConstraintSolverImpl
 {
